requirements.txt
# FastMCP - Model Context Protocol framework
fastmcp>=0.1.0

# HTTP client for Mealie API
httpx>=0.27.0

# Environment variable management
python-dotenv>=1.0.0

# Type hints
pydantic>=2.0.0
pytest-asyncio>=0.21.0
